minor refactory
